# openapi.yaml
openapi: 3.0.1
info:
  title: Queuing
  description: |-
    Queuing API 사용 방법과 예시를 담은 문서예요.

    API 사용 중 오류가 발생하거나, 질문이 있다면 새벽이라도 부담스러워 하지 마시고 연락주세요!
  version: 1.0.0
...
paths:
...
  /api/v1/user-profiles/me:
    get:
      summary: 내 프로필 조회
      deprecated: false
      description: ''
      tags:
        - 회원
      parameters: []
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponseBody'
              example:
                result:
                  nickname: 멋진사용자
                  slug: user-123abc45
                  profileImageUrl: https://example.com/profile.jpg
          headers: {}
          x-apidog-ordering: 0
        '404':
          description: 사용자 정보를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailedResponseBody'
              examples:
                UserNotFound:
                  summary: UserNotFound
                  value:
                    error:
                      statusCode: 404
                      code: user.not-found
                      message: 사용자 정보가 존재하지 않아요.
          headers: {}
          x-apidog-ordering: 1
      security: []
      x-apidog-folder: 루트/회원
      x-apidog-status: released
      x-run-in-apidog: https://app.apidog.com/web/project/1164289/apis/api-26074168-run
    patch:
      summary: 내 프로필 수정
      deprecated: false
      description: ''
      tags:
        - 회원
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponseBody'
              example:
                result: true
          headers: {}
          x-apidog-ordering: 0
        '400':
          description: 입력 값이 잘못됨 (닉네임 유효성 검사 실패 등)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailedResponseBody'
              examples:
                InvalidInput:
                  summary: InvalidInput
                  value:
                    error:
                      statusCode: 400
                      code: common.invalid-input
                      message: 입력 값이 잘못되었어요.
                      fieldErrors:
                        - field: nickname
                          value: a
                          reason: 닉네임은 2 ~ 20자 사이여야 해요.
          headers: {}
          x-apidog-ordering: 1
        '404':
          description: 사용자 정보를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailedResponseBody'
              examples:
                UserNotFound:
                  summary: UserNotFound
                  value:
                    error:
                      statusCode: 404
                      code: user.not-found
                      message: 사용자 정보가 존재하지 않아요.
          headers: {}
          x-apidog-ordering: 2
        '409':
          description: 이미 사용 중인 닉네임
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailedResponseBody'
              examples:
                NicknameDuplicated:
                  summary: NicknameDuplicated
                  value:
                    error:
                      statusCode: 409
                      code: user.nickname-duplicated
                      message: 이미 사용 중인 닉네임이에요.
          headers: {}
          x-apidog-ordering: 3
      security: []
      x-apidog-folder: 루트/회원
      x-apidog-status: released
      x-run-in-apidog: https://app.apidog.com/web/project/1164289/apis/api-26074169-run
  /api/v1/user-profiles/{slug}:
    get:
      summary: 다른 회원 프로필 조회
      deprecated: false
      description: ''
      tags:
        - 회원
      parameters:
        - name: slug
          in: path
          description: 조회할 사용자의 슬러그를 입력하세요.
          required: true
          example: ''
          schema:
            type: string
            example: user-uuid-1234
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponseBody'
              example:
                result:
                  nickname: 친구닉네임
                  slug: friend-uuid-5678
                  profileImageUrl: null
          headers: {}
          x-apidog-ordering: 0
        '404':
          description: 사용자 정보를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailedResponseBody'
              examples:
                UserNotFound:
                  summary: UserNotFound
                  value:
                    error:
                      statusCode: 404
                      code: user.not-found
                      message: 사용자 정보가 존재하지 않아요.
          headers: {}
          x-apidog-ordering: 1
      security: []
      x-apidog-folder: 루트/회원
      x-apidog-status: released
      x-run-in-apidog: https://app.apidog.com/web/project/1164289/apis/api-26074170-run
...
components:
  schemas:
    BooleanResponseBody:
      type: object
      properties:
        result:
          type: boolean
          description: 성공 여부
          example: true
      x-apidog-orders:
        - result
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
    FailedResponseBody:
      type: object
      properties:
        error:
          type: object
          properties:
            statusCode:
              type: integer
              description: HTTP 상태 코드
              example: 400
            code:
              type: string
              description: 에러 코드
              example: common.invalid-input
            message:
              type: string
              description: 에러 메시지
              example: 입력 값이 잘못되었어요.
            fieldErrors:
              type: array
              description: 필드별 검증 오류 목록 (Validation 실패 시)
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: 오류가 발생한 필드명
                  value:
                    type: string
                    description: 거부된 값
                  reason:
                    type: string
                    description: 오류 원인
                x-apidog-orders:
                  - field
                  - value
                  - reason
                x-apidog-ignore-properties: []
          x-apidog-orders:
            - statusCode
            - code
            - message
            - fieldErrors
          x-apidog-ignore-properties: []
      x-apidog-orders:
        - error
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
    UserProfileResponseBody:
      type: object
      properties:
        result:
          $ref: '#/components/schemas/UserProfileResponse'
      x-apidog-orders:
        - result
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
...
    UserProfileResponse:
      type: object
      properties:
        nickname:
          type: string
          description: 사용자 닉네임
          example: 멋진사용자
        slug:
          type: string
          description: 사용자 고유 식별자 (Slug)
          example: user-123abc45
        profileImageUrl:
          type: string
          description: 프로필 이미지 URL
          nullable: true
          example: https://example.com/profile.jpg
      x-apidog-orders:
        - nickname
        - slug
        - profileImageUrl
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
    UpdateUserProfileRequest:
      type: object
      required:
        - nickname
      properties:
        nickname:
          type: string
          minLength: 2
          maxLength: 20
          description: 변경할 닉네임 (2~20자)
          example: 새로운닉네임
      x-apidog-orders:
        - nickname
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
...
  securitySchemes: {}
servers:
  - url: https://api.queuing.patulus.com
    description: 개발
security: []